wip: internalize FormBuilder and move to module
