+++ umap/static/umap/js/modules/browser.js
@@ -4,6 +4,7 @@ import * as Icon from './rendering/icon.js'
 import * as Utils from './utils.js'
 import { EXPORT_FORMATS } from './formatter.js'
 import ContextMenu from './ui/contextmenu.js'
+import { Form } from './form/builder.js'
 
 export default class Browser {
   constructor(umap, leafletMap) {
@@ -179,7 +180,7 @@ export default class Browser {
       ],
       ['options.inBbox', { handler: 'Switch', label: translate('Current map view') }],
     ]
-    const builder = new L.FormBuilder(this, fields, {
+    const builder = new Form(this, fields, {
       callback: () => this.onFormChange(),
     })
     let filtersBuilder
@@ -189,7 +190,7 @@ export default class Browser {
     })
     if (this._umap.properties.facetKey) {
       fields = this._umap.facets.build()
-      filtersBuilder = new L.FormBuilder(this._umap.facets, fields, {
+      filtersBuilder = new Form(this._umap.facets, fields, {
         callback: () => this.onFormChange(),
       })
       DomEvent.on(filtersBuilder.form, 'reset', () => {
